Criando modulo para recursos da lista negra.
